Remove regfile from cpu.v
Instead of managing the register file from cpu.v this commit adds
inputs and output to the module that must be wired up to a regfile.
This allows narvie to use a uart transmitting regfile.

This commit is similar to 617f43e05889d4582d30cb45cf197e1bfe2ffc6b